+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * MetaProcessor.include_extra to add support for arbitrary extra values to be
   set on the meta record.
+* Correctly handled quoted svcparams when parsing SVCB/HTTPS rdata text
 
 ## v1.9.1 - 2024-06-21 - What's in a name